cars/requirements.txt
asgiref==3.3.1
Cerberus==1.3.2
Django==3.1.5
djangorestframework==3.12.2
jsonfield==3.1.0
mysql-connector-python==8.0.23
protobuf==3.14.0
PyMySQL==1.0.2
python-decouple==3.4
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
